feat: refactor RecentlyUpdatedTitles component and add TitleCard for improved structure
