+++ app/components/RecentlyUpdatedTitles/title.ts
@@ -1,0 +1,13 @@
+import { Prisma } from '@prisma/client'
+
+export type Title = Prisma.TitleGetPayload<{
+  include: {
+    Chapter: {
+      select: {
+        id: true
+        name: true
+        index: true
+      }
+    }
+  }
+}>
