+++ app/page.tsx
@@ -1,5 +1,5 @@
 import { Box } from '@chakra-ui/react'
-import RecentlyUpdatedTitles from './components/RecentlyUpdatedTitles'
+import RecentlyUpdatedTitles from './components/RecentlyUpdatedTitles/RecentlyUpdatedTitles'
 
 const Home = () => {
   return (
